new public dir structure
