+++ public/js/components/leaderboard/leaderboard.js
@@ -1,0 +1,10 @@
+const leaderBoardTemplate = require('./leaderboard.pug');
+
+const root = document.getElementById('root');
+
+
+export default function createLeaderboard() {
+  root.innerHTML = leaderBoardTemplate();
+
+  console.log('leaderboard block created');
+}
