+++ public/js/components/leaderboard/leaderboard.pug
@@ -14,4 +14,6 @@
         td=player.lvl
         td=player.score
 
-  a.button#menuButton Назад
+  a.button(
+    datahref='menu'
+  ) Назад
