+++ public/js/components/login/login.pug
@@ -1,0 +1,17 @@
+#login.block.form
+  div
+    input(
+      name='login'
+      type='text'
+      placeholder='username'
+      minlength="3"
+      required
+    )
+    input(
+      name='password'
+      type='password'
+      placeholder='password'
+    )
+    a.button(datahref='menu') login
+    p.message Not registered?
+      a(datahref='register') Create an account
