--- public/js/components/menu/login/login.pug
@@ -1,7 +1,0 @@
-#login.block.form
-  div
-    input(name='login' type='text', placeholder='username', minlength="3", required)
-    input(name='password' type='password', placeholder='password')
-    a.button#loginButton login
-    p.message Not registered?
-      a#registerButton  Create an account
