--- public/js/components/menu/settings/settings.js
@@ -1,12 +1,0 @@
-import createMenu from '../menu/menu';
-import addButtonListener from '../addbuttonlistener';
-
-import settingsTemplate from './settings.pug';
-
-
-export default function createSettings() {
-  root.innerHTML = settingsTemplate();
-  addButtonListener('menuButton', createMenu);
-
-  console.log('settings block created');
-}
