+++ public/js/components/profile/profile.js
@@ -1,11 +1,10 @@
-import createMenu from '../menu/menu';
-import addButtonListener from '../addbuttonlistener';
-
 const profileTemplate = require('./profile.pug');
+
+const root = document.getElementById('root');
+
 
 export default function createProfile() {
   root.innerHTML = profileTemplate();
-  addButtonListener('menuButton', createMenu);
 
   console.log('profile block created');
 }
