+++ public/js/components/profile/profile.pug
@@ -2,4 +2,6 @@
 .block#profile
   img(src = 'img/dieter_rams.jpg')
   h3= userName
-  a.button#menuButton Назад
+  a.button(
+    datahref='menu'
+  ) Назад
