+++ public/js/components/register/register.js
@@ -1,0 +1,12 @@
+const registerTemplate = require('./register.pug');
+
+const root = document.getElementById('root');
+
+
+export default function createRegister() {
+  root.innerHTML = registerTemplate();
+
+  validate();
+
+  console.log('register block created');
+}
