+++ public/js/components/register/register.pug
@@ -14,4 +14,6 @@
 
     a.button#registerButton create
     p.message Already registered?
-      a#loginButton  Sign In
+      a(
+        datahref='login'
+      ) Sign In
