+++ public/js/components/settings/settings.pug
@@ -3,4 +3,6 @@
   div
     input(type="checkbox", id="scales", name="feature", value="scales", checked)
     label(for='scales')
-    a.button#menuButton Назад
+    a.button(
+      datahref='menu'
+    ) Назад
